fix: base layer can be moved while measuring distances
